+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## new
 * change: Sequencex.flattenSequenceOfIterable() rename to flattenOfIterable
+* fix: Fix the stack overflow bug when Sequencex.flatten() is passed into TransformingSequence
 
 ## v1.0.0-beta02
 improve: Improve code